export cmd/{gitserver,searcher} main func for use as library
This is a prerequisite for allowing other packages to invoke gitserver and searcher as libraries. This commit alone does not expose a sufficiently flexible API for such use, but it will make the diff of such a change smaller in the future.
