+++ cmd/searcher/main.go
@@ -3,96 +3,11 @@
 package main
 
 import (
-	"context"
-	"io"
-	"log"
-	"net"
-	"net/http"
-	"os"
-	"os/signal"
-	"path/filepath"
-	"strconv"
-	"time"
-
-	"github.com/inconshreveable/log15"
-
-	"github.com/sourcegraph/sourcegraph/cmd/searcher/search"
-	"github.com/sourcegraph/sourcegraph/internal/api"
+	"github.com/sourcegraph/sourcegraph/cmd/searcher/shared"
 	"github.com/sourcegraph/sourcegraph/internal/debugserver"
-	"github.com/sourcegraph/sourcegraph/internal/env"
-	"github.com/sourcegraph/sourcegraph/internal/gitserver"
-	"github.com/sourcegraph/sourcegraph/internal/store"
-	"github.com/sourcegraph/sourcegraph/internal/trace/ot"
-	"github.com/sourcegraph/sourcegraph/internal/tracer"
 )
 
-var cacheDir = env.Get("CACHE_DIR", "/tmp", "directory to store cached archives.")
-var cacheSizeMB = env.Get("SEARCHER_CACHE_SIZE_MB", "100000", "maximum size of the on disk cache in megabytes")
-
-const port = "3181"
-
 func main() {
-	env.Lock()
-	env.HandleHelpFlag()
-	log.SetFlags(0)
-	tracer.Init()
-
 	go debugserver.Start()
-
-	var cacheSizeBytes int64
-	if i, err := strconv.ParseInt(cacheSizeMB, 10, 64); err != nil {
-		log.Fatalf("invalid int %q for SEARCHER_CACHE_SIZE_MB: %s", cacheSizeMB, err)
-	} else {
-		cacheSizeBytes = i * 1000 * 1000
-	}
-
-	service := &search.Service{
-		Store: &store.Store{
-			FetchTar: func(ctx context.Context, repo gitserver.Repo, commit api.CommitID) (io.ReadCloser, error) {
-				return gitserver.DefaultClient.Archive(ctx, repo, gitserver.ArchiveOptions{Treeish: string(commit), Format: "tar"})
-			},
-			Path:              filepath.Join(cacheDir, "searcher-archives"),
-			MaxCacheSizeBytes: cacheSizeBytes,
-		},
-		Log: log15.Root(),
-	}
-	service.Store.SetMaxConcurrentFetchTar(10)
-	service.Store.Start()
-	handler := ot.Middleware(service)
-
-	host := ""
-	if env.InsecureDev {
-		host = "127.0.0.1"
-	}
-	addr := net.JoinHostPort(host, port)
-	server := &http.Server{
-		Addr: addr,
-		Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			// For cluster liveness and readiness probes
-			if r.URL.Path == "/healthz" {
-				w.WriteHeader(200)
-				_, _ = w.Write([]byte("ok"))
-				return
-			}
-			handler.ServeHTTP(w, r)
-		}),
-	}
-	go shutdownOnSIGINT(server)
-
-	log15.Info("searcher: listening", "addr", server.Addr)
-	if err := server.ListenAndServe(); err != http.ErrServerClosed {
-		log.Fatal(err)
-	}
-}
-
-func shutdownOnSIGINT(s *http.Server) {
-	c := make(chan os.Signal, 1)
-	signal.Notify(c, os.Interrupt)
-	<-c
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	err := s.Shutdown(ctx)
-	if err != nil {
-		log.Fatal("graceful server shutdown failed, will exit:", err)
-	}
+	shared.Main()
 }
